refactor: create <oysters-brand-editor />
* add @/* as alias path
