+++ sites/src/app-root.ts
@@ -1,60 +1,13 @@
-import {
-  customElement,
-  html,
-  LitElement,
-  property,
-  query,
-  svg,
-} from 'lit-element';
-import { unsafeSVG } from 'lit-html/directives/unsafe-svg';
-import { apply, create, cssomSheet } from 'twind';
+import '@/components/oysters-brand-editor'
 
-import OystersIcon from '~/svg/oysters-icon-space.svg';
-import OystersLogo from '~/svg/oysters-logo-space.svg';
+import { customElement, html, LitElement } from 'lit-element'
+import { setup } from 'twind'
 
-import { download } from './utils/download';
-import { svg2png } from './utils/svg2png';
-
-const sheet = cssomSheet({ target: new CSSStyleSheet() });
-const { tw } = create({ sheet });
-
-const styles = {
-  button: apply`bg-gray-500 hover:bg-gray-700 text-white py-2 px-4 rounded`,
-};
+setup({ preflight: true })
 
 @customElement('app-root')
 export class AppRoot extends LitElement {
-  @property() brand = {
-    logo: OystersLogo,
-    icon: OystersIcon,
-  };
-
-  @property() type = 'logo' as 'logo' | 'icon';
-  @query('#canvas svg') svgEl!: SVGSVGElement;
-
-  static get styles() {
-    return [sheet.target];
-  }
-
-  private async download(e: Event) {
-    e.preventDefault();
-    const url = await svg2png(this.svgEl);
-    const filename = `oysters-${this.type}-space`;
-
-    download({ url, filename });
-  }
-
   render() {
-    return html`
-      <div class="${tw`flex`}" id="canvas">
-        ${svg`${unsafeSVG(this.brand[this.type])}`}
-      </div>
-      <button
-        class="${tw(styles.button)}"
-        @click="${(e: Event) => this.download(e)}"
-      >
-        Download PNG
-      </button>
-    `;
+    return html`<oysters-brand-editor></oysters-brand-editor>`
   }
 }
