+++ sites/src/components/oysters-brand-editor/index.ts
@@ -1,0 +1,1 @@
+import './oysters-brand-editor'
