+++ sites/src/components/oysters-brand-editor/styles.ts
@@ -1,0 +1,5 @@
+import { apply } from 'twind'
+
+export const styles = {
+  button: apply`bg-gray-500 hover:bg-gray-700 text-white py-2 px-4 rounded`,
+}
