+++ sites/tsconfig.json
@@ -8,11 +8,12 @@
     "baseUrl": "./",
     /* paths - import rewriting/resolving */
     "paths": {
-      "~/svg/*": ["../svg/"]
       // If you configured any Snowpack aliases, add them here.
       // Add this line to get types for streaming imports (packageOptions.source="remote"):
       //     "*": [".snowpack/types/*"]
       // More info: https://www.snowpack.dev/guides/streaming-imports
+      "~/svg/*": ["../svg/*"],
+      "@/*": ["src/*"]
     },
     /* noEmit - Snowpack builds (emits) files, not tsc. */
     "noEmit": true,
